Rework dataclass construction to support command specific columns
In order to support the Count command, the dynamic data classes need to
extract the attributes from the command.

The ColumnDefinitions are moved to the SQLCommand and are overridden
by Select (which supports joins) and Count (which adds a custom column count)

Dataclass construction itself is no longer recursive, but is built from the
column definitions at once.